Conclusão da Issue #77
