+++ compilacao/templatetags/compilacao_filters.py
@@ -156,4 +156,9 @@ def nomenclatura_heranca(d, ignore_ultimo=0, ignore_primeiro=0):
 
 @register.simple_tag
 def verbose_name(instance, field_name):
-    return instance._meta.get_field(field_name).verbose_name.title()
+    return instance._meta.get_field(field_name).verbose_name
+
+
+@register.filter
+def urldetail_content_type(obj):
+    return '%s:detail' % obj.content_type.model
