+++ static/styles/app.scss
@@ -25,8 +25,8 @@ $top-bar-dropdown-radius: $global-radius;
 	}
 }
 
-// Restyles to foundation top bar menu, in order to behave aesthetically different 
-/* 
+// Restyles to foundation top bar menu, in order to behave aesthetically different
+/*
    XXX
    Is there a better way to implement these styles using best practices without the
    need to override this much of foundation? Or at least avoid too much nesting and
@@ -38,7 +38,7 @@ $top-bar-dropdown-radius: $global-radius;
 		.has-dropdown {
 			// Default values for dropdowns, while also being hidden.
 			.dropdown {
-				// Border-radius for the first and last dropdown items. We don't style the parent dropdown, as the items overflows it. 
+				// Border-radius for the first and last dropdown items. We don't style the parent dropdown, as the items overflows it.
 				@include radius(rem-calc($top-bar-dropdown-radius));
 				// Second child here actually targets the first item, since Foundation JS injects a hidden li before it for a mobile back button.
 				> li:nth-child(2), > li:nth-child(2) > a {
@@ -50,8 +50,8 @@ $top-bar-dropdown-radius: $global-radius;
 				box-shadow: 0 10px 18px rgba(0, 0, 0, 0.19),
 							0 2px 6px rgba(0, 0, 0, 0.23);
 				// For the transition effect.
-				opacity: 0; 
-				// Show the dropdown accurately while it animates. 
+				opacity: 0;
+				// Show the dropdown accurately while it animates.
 				width: auto;
 				// This will allow the triangle pip to be visible above the dropdown.
 				overflow: visible;
@@ -68,14 +68,14 @@ $top-bar-dropdown-radius: $global-radius;
 					position: absolute;
 					top: rem-calc(-12px);
 					left: rem-calc(15px);
-				}	
-				// This bridges the gap between the top bar and a dropdown. 
+				}
+				// This bridges the gap between the top bar and a dropdown.
 				&::after {
 					content: "";
 					position: absolute;
 					z-index: -1;
 					left: 0;
-					top: rem-calc(-25px);		
+					top: rem-calc(-25px);
 					height: rem-calc(25px);
 					width: 100%;
 					// This transition is for hover-on.
@@ -94,7 +94,7 @@ $top-bar-dropdown-radius: $global-radius;
 					pointer-events: auto;
 					// Animating with a beautiful cubic-bezier curve, or Google's "Swift out" easing :)
 					transition: transform 0.3s cubic-bezier(0.55,0,0.1,1),
-								opacity 0.3s cubic-bezier(0.55,0,0.1,1), 
+								opacity 0.3s cubic-bezier(0.55,0,0.1,1),
 								// Here we make sure the clipping is set before any other transition.
 								clip 0s 0s;
 					// Don't forget to properly animate our bridge, so it keeps only between our gap.
@@ -106,6 +106,8 @@ $top-bar-dropdown-radius: $global-radius;
 		}
 	}
 }
+
+
 
 // Our app
 // - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -158,7 +160,7 @@ color: $primary-color;
 .fadein {
 -webkit-animation: fadeIn 0.25s ease-in-out;
 -moz-animation: fadeIn 0.25s ease-in-out;
--o-animation: fadeIn 0.25s ease-in-out;    
+-o-animation: fadeIn 0.25s ease-in-out;
 }
 
 .container {
@@ -181,7 +183,7 @@ color: #444444;
     display: table-row;
     box-sizing: content-box;
     }
-        
+
         /* XXX find a better way to fix main layout box-sizing */
         .page__row > .container > *{
         box-sizing: border-box;
@@ -195,6 +197,12 @@ box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 1px 4px rgba(0, 0, 0, 0.23);
     padding: 0px;
     max-width: 960px;
     }
+
+.sub-nav {
+	dt a, dd a, li a {
+		color: white !important;
+	}
+}
 
 
 .icon-bar {
@@ -301,7 +309,7 @@ text-align: center;
 
     .footer__block--about {
     padding-left: 0;
-    max-width: 195px; 
+    max-width: 195px;
     }
 
     .footer__block--license {
