+++ templates/compilacao/text_edit.html
@@ -18,12 +18,11 @@
 {% endblock %}
 
 
-{% block sections_nav %}
+{% block sections_nav %}{{block.super}}
+        <dd><a href="{% url 'ta_text' object.pk %}" class="button secondary">{% trans 'Texto' %}</a></dd>
 {% endblock %}
 
 {% block base_content %}{{block.super}}
-    {% block actions %}{% endblock %}
-    {% block detail_content %}{% endblock %}
 
 
 <div id="message_block"><div id="msg">{% trans 'Aguarde... Atualizando informações!!!'%}</div></div>
