+++ templates/compilacao/text_list.html
@@ -16,23 +16,14 @@
       {% endif %}
 {% endblock %}
 
-
-{% block title%}
-    <h1>{{ view.get_ta }}</h1>
-{% endblock %}
-
-
 {% block sections_nav %}{{block.super}}
-    <dd><a href="{% url 'ta_text_edit' object.pk %}" class="button secondary">{% trans 'Edição do Texto' %}</a></dd>
+    <dd><a href="{% url 'ta_text_edit' object.pk %}" class="button secondary">{% trans 'Edição' %}</a></dd>
 {% endblock %}
 
 {% block base_content %}{{block.super}}
     {% block actions %}{% endblock %}
-    {% block detail_content %}{% endblock %}
-
 
 <div class="cp">
-
     {% if object_list %}
 	    <div style="float: right; clear:right;">
 	    <a id="btn_font_menos" title="Diminuir tamanho da letra">a</a>
