+++ templates/materia/materialegislativa_detail.html
@@ -12,6 +12,13 @@
     <dd><a href="{% url 'numeracao' materialegislativa.id %}" class="button secondary">{% trans 'Numeração' %}</a></dd>
     <dd><a href="{% url 'tramitacao_materia' materialegislativa.id %}" class="button secondary">{% trans 'Tramitação' %}</a></dd>
     <dd><a href="{% url 'relatoria' materialegislativa.id %}" class="button secondary">{% trans 'Relatoria' %}</a></dd>
+
+    {% comment %}
+        Opção adicionada para chamar o TextoArticulado da matéria.
+            para integração foram necessárias apenas criar a url materia_ta em urls.py
+            e a view MateriaTaView(IntegracaoTaView) em views.py
+           Em nada mais a integração interfere em MateriaLegislativa
+    {% endcomment %}
     <dd><a href="{% url 'materia_ta' materialegislativa.id %}" class="button secondary">{% trans 'Texto' %}</a></dd>
   </dl>
 {% endblock sections_nav %}
