+++ templates/norma/normajuridica_detail.html
@@ -1,15 +1,16 @@
 {% extends "crud/detail.html" %}
 {% load i18n %}
-{% block actions %}
-<dl class="sub-nav right">
-    <dd><a href="{% url 'ta_edit' object.pk %}" class="button">{% trans 'Editar' %}</a></dd>
-    <dd><a href="{% url 'ta_delete' object.pk %}" class="button alert">{% trans 'Excluir' %}</a></dd>
-</dl>
-{% endblock actions %} 
 {% block sections_nav %}
     <dl class="sub-nav left">
         <dd><a href="{% url 'normajuridica:detail' object.pk %}" class="button secondary">{% trans 'Início' %}</a></dd>
+
+        {% comment %}
+            Opção adicionada para chamar o TextoArticulado da norma.
+                para integração foram necessárias apenas criar a url norma_ta em urls.py
+                e a view NormaTa(IntegracaoTaView) em views.py
+               Em nada mais a integração interfere em Norma Jurídica
+        {% endcomment %}
         <dd><a href="{% url 'norma_ta' object.pk %}" class="button secondary">{% trans 'Texto' %}</a></dd>
     </dl>
-        
+
 {% endblock sections_nav %}
