+++ templates/sistema.html
@@ -77,16 +77,6 @@
   		<li><a href="">Assunto de Norma Jurídica</a></li>
 	</ul>
 
-	<h5>Módulo Compilação</h5>
-	<ul class="small-block-grid-1 medium-block-grid-2 large-block-grid-2">
-  		<li><a href="{% url 'tipodispositivo:list' %}">Tipo de Dispositivo</a></li> 
-  		<li><a href="{% url 'tiponota:list' %}">Tipo de Nota</a></li> 
-  		<li><a href="{% url 'tipovide:list' %}">Tipo de Vide</a></li> 
-  		<li><a href="{% url 'tipopublicacao:list' %}">Tipo de Publicação</a></li> 
-        <li><a href="{% url 'veiculopublicacao:list' %}">Veículo de Públicação</a></li> 
-        <li><a href="{% url 'perfilestruturaltextosnormativos:list' %}">Perfil Estrutural de Textos Normativos</a></li> 
-	</ul>
-
 	<h5>Módulo Sessão Plenária</h5>
 	<ul class="small-block-grid-1 medium-block-grid-2 large-block-grid-2">
   		<li><a href="{% url 'tiposessaoplenaria:list' %}">Tipo de Sessão Plenária</a></li>
